+++ includes/functions/parser.h
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 20:41:18 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/28 22:21:45 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/04/15 20:18:57 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,9 @@ void		get_items(t_item **list, char *argv);
 t_corners	set_corners(t_item item);
 t_coords	**set_cam_vectors(t_corners corn);
 void		free_array_coords(t_coords ***list);
-
+int			try_set_atof(float *out, char *nbr, long min, long max);
+int			try_set_color(t_colors *out, const char *rgb);
+int			try_set_coords(t_coords *out, const char *coords, long min, \
+				long max);
 
 #endif