+++ includes/options.h
@@ -1,31 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minirt.h                                           :+:      :+:    :+:   */
+/*   options.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/08 20:57:46 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/04/15 20:49:50 by emadriga         ###   ########.fr       */
+/*   Created: 2022/04/15 15:19:40 by emadriga          #+#    #+#             */
+/*   Updated: 2022/04/15 15:29:24 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef MINIRT_H
-# define MINIRT_H
-# include <unistd.h>
-# include <stdlib.h>
-# include <string.h>
-# include <stdio.h>
-# include <fcntl.h>
-# include <mlx.h>
-# include <math.h>
-# include "libft.h"
+#ifndef OPTIONS_H
+# define OPTIONS_H
 # include "constants.h"
-# include "dtos.h"
-# include "options.h"
-# include "functions/coloring.h"
-# include "functions/parser.h"
-# include "functions/render.h"
-# include "functions/utils.h"
+# define DEFAULT_LIGHT_COLOR TRUE
 
 #endif