+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+## 2022-04-15 - Edu
+### Added
+- Added **try_parse** to handle parser for mandatory items.
 ## 2022-04-10 - Edu
 ### Changed
 - Document **render**